fixed code so that Search component uses the same ReservationsTable component as Dashboard
