+++ front-end/src/dashboard/Search.js
@@ -1,6 +1,6 @@
-import { findReservationsByPhoneNumber } from "../utils/api";
+import { listReservations } from "../utils/api";
 import { useState, useEffect } from "react";
-import { Link } from "react-router-dom"
+import ReservationsTable from "./ReservationsTable";
 import ErrorAlert from "../layout/ErrorAlert";
 
 
@@ -12,7 +12,6 @@ const Search = () => {
   const [reservations, setReservations] = useState([]);
   const [reservationsError, setReservationsError] = useState(null)
 
-  const noReservationsError = { message: "No Reservations Found."}
 
 
 
@@ -52,48 +51,48 @@ const Search = () => {
 //   }
 
 
-function reservationsByPhoneNumber(reservations) {
-  return  (
-      <div className="row g-2">
-        <table className="table table-dark">
-  <thead>
-    <tr>
-      <th scope="col">Last Name</th>
-      <th scope="col">First Name</th>
-      <th scope="col">Mobile Number</th>
-      <th scope="col">Time</th>
-      <th scope="col">Party Of</th>
-      <th scope="col">Status</th>
-    </tr>
-  </thead>
-  <tbody>
-  {
-    reservations
-    .filter((reservation) => reservation.mobile_number.includes(phoneNumber.mobile_number))
-    .map((reservation) => 
-  (
-            <tr key={reservation.reservation_id}>
-            <td>{reservation.last_name}</td>
-            <td>{reservation.first_name}</td>
-            <td>{reservation.mobile_number}</td>
-            <td>{reservation.reservation_time}</td>
-            <td>{reservation.people}</td>
-            <td data-reservation-id-status={reservation.reservation_id}>{reservation.status}</td>
-            {reservation.status === "booked" && <td><Link to={`/reservations/${reservation.reservation_id}/seat`} type="button" className="btn btn-success">Seat</Link></td>}
-          </tr>
-  ))}
-  </tbody>
-</table>
+// function reservationsByPhoneNumber(reservations) {
+//   return  (
+//       <div className="row g-2">
+//         <table className="table table-dark">
+//   <thead>
+//     <tr>
+//       <th scope="col">Last Name</th>
+//       <th scope="col">First Name</th>
+//       <th scope="col">Mobile Number</th>
+//       <th scope="col">Time</th>
+//       <th scope="col">Party Of</th>
+//       <th scope="col">Status</th>
+//     </tr>
+//   </thead>
+//   <tbody>
+//   {
+//     reservations
+//     .filter((reservation) => reservation.mobile_number.includes(phoneNumber.mobile_number))
+//     .map((reservation) => 
+//   (
+//             <tr key={reservation.reservation_id}>
+//             <td>{reservation.last_name}</td>
+//             <td>{reservation.first_name}</td>
+//             <td>{reservation.mobile_number}</td>
+//             <td>{reservation.reservation_time}</td>
+//             <td>{reservation.people}</td>
+//             <td data-reservation-id-status={reservation.reservation_id}>{reservation.status}</td>
+//             {reservation.status === "booked" && <td><Link to={`/reservations/${reservation.reservation_id}/seat`} type="button" className="btn btn-success">Seat</Link></td>}
+//           </tr>
+//   ))}
+//   </tbody>
+// </table>
 
-      </div>
-    )
-}
+//       </div>
+//     )
+// }
 
   function listReservationsByPhoneNumber(e) {
     e.preventDefault();
     const abortController = new AbortController();
     setReservationsError(null);
-    findReservationsByPhoneNumber(phoneNumber, abortController.signal)
+    listReservations(phoneNumber, abortController.signal)
     .then(setReservations)
     .catch(setReservationsError)
     return () => abortController.abort();
@@ -122,7 +121,10 @@ function reservationsByPhoneNumber(reservations) {
   	</button>
     </form>
     <div>
-    { reservationsByPhoneNumber(reservations) }
+      {reservations.length ? 
+    <ReservationsTable reservations={reservations} />
+    : <h1> No reservations found. </h1>
+      }
     </div>
     </div>
   );
